Feat :: create DeviceMapper #36
DeviceService, convertDevicesToDtos 함수 이동
