+++ src/main/java/com/keepgoing/keepserver/domain/device/service/DeviceService.java
@@ -1,9 +1,6 @@
 package com.keepgoing.keepserver.domain.device.service;
 
-import com.keepgoing.keepserver.domain.device.entity.Device;
-import com.keepgoing.keepserver.domain.device.entity.DeviceStatus;
 import com.keepgoing.keepserver.domain.device.payload.request.DeviceDto;
-import com.keepgoing.keepserver.domain.device.payload.response.DeviceResponseDto;
 import com.keepgoing.keepserver.global.common.BaseResponse;
 import org.springframework.security.core.Authentication;
 
@@ -19,21 +16,4 @@ public interface DeviceService {
 
     BaseResponse findAll();
     BaseResponse rentDevice(String deviceName, String email);
-
-    default DeviceResponseDto entityToDto(Device entity) {
-        return DeviceResponseDto.builder()
-                .id(entity.getId())
-                .deviceName(entity.getDeviceName())
-                .imgUrl(entity.getImgUrl())
-                .status(entity.getStatus())
-                .build();
-    }
-
-    default Device dtoToEntity(DeviceDto dto) {
-        return Device.builder()
-                .deviceName(dto.deviceName())
-                .imgUrl(dto.imgUrl())
-                .status(DeviceStatus.AVAILABLE)
-                .build();
-    }
 }
